+++ NEWS.md
@@ -1,3 +1,8 @@
+# enviGCMS 0.6.0
+
+- remove dependency of rcdk package 
+- CRAN
+
 # enviGCMS 0.5.9
 
 - fix `...` issue